Update mmc.cpp
Commented out debugging methods but left them here for future use.

Made error messages more specific, and added some other error checks.

Changed the way the layout file is read, to no longer require the number of levels to be actually in the file. The number of levels is now deduced from the number of lines read before an EOF is reached.

Fixed bug in get_next_worker that checked for workers not even under the manager.